ITM-111 & ITM-93: Break-up itm_scenario_session into different files (#26)
* Moved action validation/processing code to `ITMActionHandler`
* Refactored some action handling code
* Decoupled history JSON output from DB, and only init DB once, 
* Moved history to `ITMHistory` with its own ability to write JSON output
* After discussion, consume supplies (and advance simulated time) even if treatment doesn't treat the injury
* Removed `git_push.sh` (for ITM-93)
